Move source code into src/
