+++ src/examples/standard.js
