+++ src/lib/NodeHandle.js
