+++ src/lib/Publisher.js
