+++ src/lib/RosNode.js
