+++ src/lib/Subscriber.js
