+++ src/tools/flatten.js
