+++ src/tools/generateMessages.js
