+++ src/utils/network_utils.js
