+++ src/utils/serialization_utils.js
